refactor(dashboard): consolidate text truncation into src/lib/truncate.ts
Unify 6 duplicate truncate/trimText functions from 6 files into 2 functions:
- trimText: whitespace-collapsing, null-returning for empty input
- truncate: simple slice with ellipsis

Net: -8 lines (34 removed, 26 added). Zero behavior change.
